+++ tests/unit/Mock/PreferenceMock.php
@@ -48,7 +48,8 @@ class PreferenceMock
         "area_code" => "",
         "number" => "99999999999"
       ],
-      "surname" => "Fontes"
+      "surname" => "Fontes",
+      "is_prime_user" => true
     ],
     "payment_methods" => [
       "excluded_payment_methods" => [
@@ -119,7 +120,8 @@ class PreferenceMock
         "register_updated_at"=> "2023-04-11T19=>18=>38.486Z",
         "user_email"=> "test@test.com",
         "authentication_type"=> "",
-        "last_purchase"=> ""
+        "last_purchase"=> "",
+        "is_prime_user"=> true
       ],
       "seller"=> [
         "id"=> "",